BRS-219: upload to s3 before updating cloudfront distribution (#84)
Updating the cloudfront distribution can take 2+ minutes (up to 25 or
so), during which some edge locations will try to access the new
content, even if it hasn't been uploaded to S3 yet.

Note this creates a potential problem if we change the S3 bucket name,
since the bucket is managed by terragrunt. I haven't found a solution
for that - it shouldn't be something that happens often though.

Also format actions files with prettier.